requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
streamlit>=1.30.0
chromadb>=0.4.20
google-generativeai>=0.3.0
langchain-text-splitters>=0.0.1
folium>=0.15.0
streamlit-folium>=0.15.0
